Updated speleo_sensor for kernel length problems.
SDEE: added error messages and set the length of the transit time distribution (tau, h) to avoid the case where the convolution kernel exceeds half the length of the total time series. The code will now return a warning for this case, and sets the length of tau such that the kernel that has reached zero for all practical purposes.